fix ci

+++ tests/ut/quantization/test_w8a8.py
@@ -769,10 +769,10 @@ class TestSelectExperts(TestBase):
         mock_topk.return_value = (torch.ones(self.num_tokens, self.top_k),
                                   torch.zeros(self.num_tokens,
                                               self.top_k,
-                                              dtype=torch.long),
+                                              dtype=torch.int32),
                                   torch.arange(0,
                                                self.num_tokens * self.top_k,
-                                               dtype=torch.int32).view(
+                                               dtype=torch.float).view(
                                                    self.top_k,
                                                    -1).permute(1,
                                                                0).contiguous())
@@ -880,10 +880,10 @@ class TestSelectExperts(TestBase):
         mock_topk.return_value = (torch.ones(self.num_tokens, self.top_k),
                                   torch.zeros(self.num_tokens,
                                               self.top_k,
-                                              dtype=torch.long),
+                                              dtype=torch.int32),
                                   torch.arange(0,
                                                self.num_tokens * self.top_k,
-                                               dtype=torch.int32).view(
+                                               dtype=torch.float).view(
                                                    self.top_k,
                                                    -1).permute(1,
                                                                0).contiguous())
@@ -906,10 +906,10 @@ class TestSelectExperts(TestBase):
         mock_topk.return_value = (torch.ones(self.num_tokens, self.top_k),
                                   torch.zeros(self.num_tokens,
                                               self.top_k,
-                                              dtype=torch.long),
+                                              dtype=torch.int32),
                                   torch.arange(0,
                                                self.num_tokens * self.top_k,
-                                               dtype=torch.int32).view(
+                                               dtype=torch.float).view(
                                                    self.top_k,
                                                    -1).permute(1,
                                                                0).contiguous())
